% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_half_donut.R
\name{combine_half_donut}
\alias{combine_half_donut}
\title{Combine Two Half Donut Plots}
\usage{
combine_half_donut(
  tier_id,
  left_cover_type = "HARD CORAL",
  right_cover_type = "MACROALGAE"
)
}
\arguments{
\item{tier_id}{Character. The unique identifier for the region/tier to query.}

\item{left_cover_type}{Character. Cover type for the left half donut (default = "HARD CORAL").}

\item{right_cover_type}{Character. Cover type for the right half donut (default = "MACROALGAE").}
}
\value{
A combined \code{ggplot} object with both half-donut charts.
}
\description{
This function generates two half-donut plots (left and right) for different cover types
and combines them into a single figure.
}
\examples{
combine_half_donuts(tier_id = "exampleTierID")

}
