% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_elisa_report.R
\name{create_temp_directory}
\alias{create_temp_directory}
\title{Create Temporary Working Directory}
\usage{
create_temp_directory()
}
\value{
Path to temporary directory
}
\description{
Create Temporary Working Directory
}
\keyword{internal}
