% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_elisa_report.R
\name{generate_pdf}
\alias{generate_pdf}
\title{Generate PDF Report}
\usage{
generate_pdf(file, params, output_file)
}
\arguments{
\item{file}{Normalized path to input file}

\item{params}{List of report parameters}

\item{output_file}{Final output file path}
}
\value{
Invisible path to generated PDF
}
\description{
Generate PDF Report
}
\keyword{internal}
