% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tier_boundary.R
\name{get_tiers_frombbox}
\alias{get_tiers_frombbox}
\title{Retrieve Tier IDs and boundaries from ReefCloud API}
\usage{
get_tiers_frombbox(tier_level = 4, bbox)
}
\arguments{
\item{tier_level}{Integer. The tier level to retrieve (must be between 2 and 6, default = 4).}

\item{bbox}{A named list or vector containing bounding box coordinates.}
}
\value{
An \code{sf} object representing the tier boundaries.
}
\description{
Retrieves spatial tier boundaries from the ReefCloud API based on a specified tier level and bounding box.
}
\details{
The function gets all tier ids for a specified tier level and bounding box using the ReefCloud API.
Tier levels defaults to 4 (region) but can be from 2 (country) to 6 (site).
}
