% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_elisa_report.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate Input Parameters}
\usage{
validate_parameters(method, errorbars, point_size, linewidth, errorbar_width)
}
\description{
Validate Input Parameters
}
\keyword{internal}
